Question 2.
I followed the instruction based on what need to do with all the data and use the knowledge in the class. ex. how to do a pivot table. The main challenge was to understand what the questions are asking and what are the ending results need to be shown. I read the questions multiple times and use my rational to judge what are the key information.